backend/requirements.txt
fastapi
pymongo
guidance
requests
openai
python-dotenv
uvicorn
pydantic
pillow